middleware refactored
